Move ACL logic and queries into separate ACL files
The ACL logic and queries were getting quite large and deserves its own
file at this point.
