+++ backup/dependencies_test.go
@@ -3,6 +3,7 @@ package backup_test
 import (
 	"github.com/greenplum-db/gp-common-go-libs/testhelper"
 	"github.com/greenplum-db/gpbackup/backup"
+	"github.com/greenplum-db/gpbackup/testutils"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
@@ -21,6 +22,7 @@ var _ = Describe("backup/dependencies tests", func() {
 		relation2 = backup.Relation{Schema: "public", Name: "relation2", Oid: 2}
 		relation3 = backup.Relation{Schema: "public", Name: "relation3", Oid: 3}
 		depMap = make(map[backup.UniqueID]map[backup.UniqueID]bool, 0)
+		toc, backupfile = testutils.InitializeTestTOC(buffer, "predata")
 	})
 	Describe("TopologicalSort", func() {
 		It("returns the original slice if there are no dependencies among objects", func() {
@@ -114,6 +116,135 @@ var _ = Describe("backup/dependencies tests", func() {
 				backup.UniqueID{Oid: 8}: backup.ObjectMetadata{Comment: "text search dictionary"},
 			}
 			Expect(result).To(Equal(expected))
+		})
+	})
+	Describe("PrintDependentObjectStatements", func() {
+		var (
+			objects      []backup.Sortable
+			metadataMap  backup.MetadataMap
+			tableDefsMap map[uint32]backup.TableDefinition
+			funcInfoMap  map[uint32]backup.FunctionInfo
+		)
+		BeforeEach(func() {
+			funcInfoMap = map[uint32]backup.FunctionInfo{
+				1: {QualifiedName: "public.write_to_s3", Arguments: "", IsInternal: false},
+				2: {QualifiedName: "public.read_from_s3", Arguments: "", IsInternal: false},
+			}
+			objects = []backup.Sortable{
+				backup.Function{Oid: 1, Schema: "public", Name: "function", FunctionBody: "SELECT $1 + $2",
+					Arguments: "integer, integer", IdentArgs: "integer, integer", ResultType: "integer", Language: "sql"},
+				backup.Type{Oid: 2, Schema: "public", Name: "base", Type: "b", Input: "typin", Output: "typout", Category: "U"},
+				backup.Type{Oid: 3, Schema: "public", Name: "composite", Type: "c", Attributes: []backup.Attribute{{Name: "foo", Type: "integer"}}, Category: "U"},
+				backup.Type{Oid: 4, Schema: "public", Name: "domain", Type: "d", BaseType: "numeric", Category: "U"},
+				backup.Relation{Oid: 5, Schema: "public", Name: "relation"},
+				backup.ExternalProtocol{Oid: 6, Name: "ext_protocol", Trusted: true, ReadFunction: 2, WriteFunction: 1, Validator: 0},
+			}
+			metadataMap = backup.MetadataMap{
+				backup.UniqueID{ClassID: backup.PG_PROC_OID, Oid: 1}:        backup.ObjectMetadata{Comment: "function"},
+				backup.UniqueID{ClassID: backup.PG_TYPE_OID, Oid: 2}:        backup.ObjectMetadata{Comment: "base type"},
+				backup.UniqueID{ClassID: backup.PG_TYPE_OID, Oid: 3}:        backup.ObjectMetadata{Comment: "composite type"},
+				backup.UniqueID{ClassID: backup.PG_TYPE_OID, Oid: 4}:        backup.ObjectMetadata{Comment: "domain"},
+				backup.UniqueID{ClassID: backup.PG_CLASS_OID, Oid: 5}:       backup.ObjectMetadata{Comment: "relation"},
+				backup.UniqueID{ClassID: backup.PG_EXTPROTOCOL_OID, Oid: 6}: backup.ObjectMetadata{Comment: "protocol"},
+			}
+			tableDefsMap = map[uint32]backup.TableDefinition{
+				5: {DistPolicy: "DISTRIBUTED RANDOMLY", ColumnDefs: []backup.ColumnDefinition{}},
+			}
+		})
+		It("prints create statements for dependent types, functions, protocols, and tables (domain has a constraint)", func() {
+			constraints := []backup.Constraint{
+				{Name: "check_constraint", ConDef: "CHECK (VALUE > 2)", OwningObject: "public.domain"},
+			}
+			backup.PrintDependentObjectStatements(backupfile, toc, objects, metadataMap, tableDefsMap, constraints, funcInfoMap)
+			testhelper.ExpectRegexp(buffer, `
+CREATE FUNCTION public.function(integer, integer) RETURNS integer AS
+$_$SELECT $1 + $2$_$
+LANGUAGE sql;
+
+
+COMMENT ON FUNCTION public.function(integer, integer) IS 'function';
+
+
+CREATE TYPE public.base (
+	INPUT = typin,
+	OUTPUT = typout
+);
+
+
+COMMENT ON TYPE public.base IS 'base type';
+
+
+CREATE TYPE public.composite AS (
+	foo integer
+);
+
+COMMENT ON TYPE public.composite IS 'composite type';
+
+CREATE DOMAIN public.domain AS numeric
+	CONSTRAINT check_constraint CHECK (VALUE > 2);
+
+
+COMMENT ON DOMAIN public.domain IS 'domain';
+
+
+CREATE TABLE public.relation (
+) DISTRIBUTED RANDOMLY;
+
+
+COMMENT ON TABLE public.relation IS 'relation';
+
+
+CREATE TRUSTED PROTOCOL ext_protocol (readfunc = public.read_from_s3, writefunc = public.write_to_s3);
+
+
+COMMENT ON PROTOCOL ext_protocol IS 'protocol';
+`)
+		})
+		It("prints create statements for dependent types, functions, protocols, and tables (no domain constraint)", func() {
+			constraints := []backup.Constraint{}
+			backup.PrintDependentObjectStatements(backupfile, toc, objects, metadataMap, tableDefsMap, constraints, funcInfoMap)
+			testhelper.ExpectRegexp(buffer, `
+CREATE FUNCTION public.function(integer, integer) RETURNS integer AS
+$_$SELECT $1 + $2$_$
+LANGUAGE sql;
+
+
+COMMENT ON FUNCTION public.function(integer, integer) IS 'function';
+
+
+CREATE TYPE public.base (
+	INPUT = typin,
+	OUTPUT = typout
+);
+
+
+COMMENT ON TYPE public.base IS 'base type';
+
+
+CREATE TYPE public.composite AS (
+	foo integer
+);
+
+COMMENT ON TYPE public.composite IS 'composite type';
+
+CREATE DOMAIN public.domain AS numeric;
+
+
+COMMENT ON DOMAIN public.domain IS 'domain';
+
+
+CREATE TABLE public.relation (
+) DISTRIBUTED RANDOMLY;
+
+
+COMMENT ON TABLE public.relation IS 'relation';
+
+
+CREATE TRUSTED PROTOCOL ext_protocol (readfunc = public.read_from_s3, writefunc = public.write_to_s3);
+
+
+COMMENT ON PROTOCOL ext_protocol IS 'protocol';
+`)
 		})
 	})
 })
